fix: move `patch_integration_settings` to `apps/workspaces/actions.py`
Avoids a circular import between actions.py and tasks.py
