+++ apps/workspaces/tasks.py
@@ -6,7 +6,7 @@ from fyle_integrations_platform_connector import PlatformConnector
 from fyle_rest_auth.helpers import get_fyle_admin
 
 from apps.fyle.enums import FundSourceEnum
-from apps.fyle.helpers import patch_request, post_request
+from apps.fyle.helpers import post_request
 from apps.fyle.tasks import async_create_expense_groups
 from apps.mappings.models import TenantMapping
 from apps.tasks.enums import TaskLogStatusEnum, TaskLogTypeEnum
@@ -153,26 +153,3 @@ def post_to_integration_settings(workspace_id: int, active: bool):
         post_request(url, payload, refresh_token)
     except Exception as error:
         logger.error(error)
-
-
-def patch_integration_settings(workspace_id: int, errors: int = None, is_token_expired = None):
-    """
-    Patch integration settings
-    """
-
-    refresh_token = FyleCredential.objects.get(workspace_id=workspace_id).refresh_token
-    url = '{}/integrations/'.format(settings.INTEGRATIONS_SETTINGS_API)
-    payload = {
-        'tpa_name': 'Fyle Xero Integration',
-    }
-
-    if errors is not None:
-        payload['errors_count'] = errors
-
-    if is_token_expired is not None:
-        payload['is_token_expired'] = is_token_expired
-
-    try:
-        patch_request(url, payload, refresh_token)
-    except Exception as error:
-        logger.error(error, exc_info=True)
